Split Move request specs into a file per action
